moved calculator files to lib/recipes
