+++ lib/recipes/ingredient_constants.rb
